Fixes mulitple bugs
Fixes a bug that occurs when trying to do timing analysis when no send keypress is present in the audio file being used as input to the DSP tool. Also fixes a bug that was causing the time it takes to type a word to be calcuated incorretly. This was due to improper use of the startWord and endWord indexes. Finally fixed a bug that was created when fixing the other two bugs. This bug threw an error when printing the calcuated time between each keypress when the last keypress was not detected as a send keypress.